This test harms me

+++ src/main/java/com/practice/devLit/repository/SubjectsRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.practice.devLit.model.Subjects;
 
 @Repository
-public interface SubjectsRepository extends CrudRepository<Subjects, Integer> {
+public interface SubjectsRepository extends CrudRepository<Subjects, Long> {
 
 }
